requirement.txt
pandas
numpy
scikit-learn
faiss-cpu 
openai
flask

